backend/requirements.txt
flask
faiss-cpu
sentence-transformers
gdown
numpy
requests
openai
time
json
numpy